requirements.txt
numpy>=1.19.4
PyYAML>=5.3.1
jedi==0.17.2
ipython>=7.19.0
cowsay
colorama
art